Separated Response and ResponseBuilder and added description attribute
