+++ bartbot/process/controller.py
@@ -3,7 +3,8 @@ import logging
 from abc import (ABC, abstractmethod)
 
 from bartbot.receive.message import (Message)
-from bartbot.send.response import (Response, ResponseBuilder)
+from bartbot.send.response import (Response)
+from bartbot.send.response_builder import (ResponseBuilder)
 
 
 def import_controller(controllerName: str):
@@ -28,18 +29,3 @@ class EchoController(Controller):
 
     def produce_responses(self) -> ResponseBuilder:
         return ResponseBuilder(recipientId=self.message.senderId, text=f"You typed {self.message.text if hasattr(self.message, 'text') else '[no text found]'}")
-
-
-# TODO:
-# - flesh out classes below
-
-# - media template
-#   https://developers.facebook.com/docs/messenger-platform/reference/template/media
-
-# - postback, web_url and simple share button
-#   https://developers.facebook.com/docs/messenger-platform/reference/buttons/postback
-#   https://developers.facebook.com/docs/messenger-platform/reference/buttons/url
-#   https://developers.facebook.com/docs/messenger-platform/reference/buttons/share
-
-# - complex share button
-#   https://developers.facebook.com/docs/messenger-platform/reference/buttons/share
